Exceptions handling

+++ Backend/src/main/java/ru/blss/lab1/controller/ShoppingCartController.java
@@ -6,6 +6,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.blss.lab1.domain.StoreItem;
 import ru.blss.lab1.domain.User;
+import ru.blss.lab1.exception.NoMoreItemException;
+import ru.blss.lab1.exception.UnauthorizedUserException;
 import ru.blss.lab1.service.ShoppingCartService;
 
 import javax.servlet.http.HttpServletRequest;
@@ -21,14 +23,22 @@ public class ShoppingCartController extends ApiController {
     }
 
     @PostMapping("cart/add")
-    public void addItem(HttpServletRequest request, @RequestBody StoreItem item) {
+    public void addItem(HttpServletRequest request, @RequestBody StoreItem item) throws UnauthorizedUserException, NoMoreItemException {
         User user = getUser(request);
+        if (user == null) {
+            throw new UnauthorizedUserException();
+        }
         shoppingCartService.AddItemToCart(user, item);
     }
 
     @PostMapping("cart/remove")
-    public void removeItem(HttpServletRequest request, @RequestBody StoreItem item) {
+    public void removeItem(HttpServletRequest request, @RequestBody StoreItem item) throws UnauthorizedUserException {
         User user = getUser(request);
+        if (user == null) {
+            throw new UnauthorizedUserException();
+        }
         shoppingCartService.RemoveItemFromCart(user, item);
     }
+
+    //todo get current cart
 }
